+++ MainForm.Designer.cs
@@ -121,7 +121,7 @@ namespace LockMyEthTool
             this.InitialEth2SetupButton.Name = "InitialEth2SetupButton";
             this.InitialEth2SetupButton.Size = new System.Drawing.Size(424, 23);
             this.InitialEth2SetupButton.TabIndex = 4;
-            this.InitialEth2SetupButton.Text = "Initial Eth2 Setup (Download prysm and import medalla validator keys)";
+            this.InitialEth2SetupButton.Text = "Eth2 Setup (Download executables and import medalla validator keys)";
             this.InitialEth2SetupButton.UseVisualStyleBackColor = true;
             this.InitialEth2SetupButton.Click += new System.EventHandler(this.InitialEth2SetupButton_Click);
             //